backend/requirements.txt
fastapi
uvicorn
yt-dlp
httpx
python-dotenv
cachetools
slowapi
